requirements.txt
# Python dependencies for Azure DevOps pipeline and automation
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
gitpython>=3.1.0
requests>=2.31.0

